Apply Geographically Weighted Regression (GWR)

This function applies Geographically Weighted Regression (GWR) to a given formula and dataset. It allows for bandwidth selection and adaptive bandwidth, with optional progress messages.

apply.GWR(
  formula,
  data,
  longlat = TRUE,
  cl = NULL,
  fit.points = NULL,
  showProgressSteps = FALSE
)

Arguments

formula

A formula object describing the model to be fitted.

data

A spatial data frame containing the variables in the formula.

longlat

Logical; if TRUE, coordinates are treated as longitude/latitude.

cl

Optional cluster object for parallel processing.

fit.points

Optional spatial data frame for making predictions.

showProgressSteps

Logical; if TRUE, prints progress messages.

Value

A list containing the inputs and the GWR output.

Examples

data(meuse.grid, package = "spEnviroDistr")
coordinates(meuse.grid) = ~x+y
proj4string(meuse.grid) <- CRS("+init=epsg:28992")
#> Warning: Discarded datum Amersfoort in Proj4 definition
gridded(meuse.grid) <- TRUE

data(meuse, package = "spEnviroDistr")
coordinates(meuse) = ~x+y
proj4string(meuse) <- CRS("+init=epsg:28992")

cl <- createCluster(free = 2)
RESULT <- apply.GWR(sqrt(copper) ~ alt + dist, meuse, longlat = FALSE, cl = cl)
#> Warning: CRS object has comment, which is lost in output; in tests, see
#> https://cran.r-project.org/web/packages/sp/vignettes/CRS_warnings.html
#> Warning: Discarded datum Unknown based on Bessel 1841 ellipsoid in Proj4 definition

print(RESULT$output)
#> Call:
#> gwr(formula = formula, data = data, bandwidth = bwG, adapt = adG, 
#>     hatmatrix = TRUE, longlat = longlat, cl = cl, predictions = TRUE)
#> Kernel function: gwr.Gauss 
#> Adaptive quantile: 0.0106526 (about 1 of 155 data points)
#> Summary of GWR coefficient estimates at data points:
#>                     Min.     1st Qu.      Median     3rd Qu.        Max.
#> X.Intercept. -53.9825901   6.8433353  20.7548716  37.5414276 105.8657675
#> alt           -2.6028035  -0.8123230  -0.3634956   0.0078075   1.5706485
#> dist         -37.4219736 -13.2348760  -5.8716219  -2.0080712  13.3767655
#>               Global
#> X.Intercept. 17.2254
#> alt          -0.2718
#> dist         -4.9757
#> Number of data points: 155 
#> Effective number of parameters (residual: 2traceS - traceS'S): 89.00475 
#> Effective degrees of freedom (residual: 2traceS - traceS'S): 65.99525 
#> Sigma (residual: 2traceS - traceS'S): 0.8370865 
#> Effective number of parameters (model: traceS): 69.26538 
#> Effective degrees of freedom (model: traceS): 85.73462 
#> Sigma (model: traceS): 0.7344269 
#> Sigma (ML): 0.5462115 
#> AICc (GWR p. 61, eq 2.33; p. 96, eq. 4.21): 512.5333 
#> AIC (GWR p. 96, eq. 4.22): 321.6641 
#> Residual sum of squares: 46.24378 
#> Quasi-global R2: 0.894203 

RESULT <- apply.GWR(log(copper) ~ alt + dist + elev, meuse, longlat = FALSE, cl = cl, fit.points = meuse.grid)
#> Warning: CRS object has comment, which is lost in output; in tests, see
#> https://cran.r-project.org/web/packages/sp/vignettes/CRS_warnings.html
#> Warning: CRS object has comment, which is lost in output; in tests, see
#> https://cran.r-project.org/web/packages/sp/vignettes/CRS_warnings.html